fix(cli): split mount preflight from exec
