+++ crates/mvm-cli/src/exec.rs
@@ -29,10 +29,12 @@ use crate::ui;
 /// Matches GNU `timeout(1)` so scripts can branch on it.
 mod guest_run;
 mod launch_plan;
+mod mounts;
 mod session;
 mod transient;
 
 pub use launch_plan::load_launch_plan;
+pub(crate) use mounts::refuse_unsupported_dir_shares;
 
 use guest_run::{emit_guest_console_diagnostic, run_in_guest, run_wasm_module};
 use session::wait_for_agent_timed;
@@ -1341,36 +1343,6 @@ pub struct ResolvedLaunch {
 /// what makes the config claim-eligible. A warm spawn passes `None`: a factory
 /// parent carries no workload authority, and the spawn drops the tenant and
 /// plan from the config it is handed anyway.
-/// Refuse `--mount` on a backend with no virtio-fs share device, naming the
-/// backend and what to use instead.
-///
-/// `--mount` is advertised without qualification but only one backend serves
-/// it, so on every other backend the flag is accepted, paid for, and then
-/// rejected mid-launch. Deciding here costs nothing and says which backend is
-/// the problem — the runner's own refusal names the volume but not the reason
-/// the caller can act on, which is *which backend they are on*.
-pub(crate) fn refuse_unsupported_dir_shares(
-    backend_name: &str,
-    supports_dir_shares: bool,
-    dir_shares: &[DirShareSpec],
-) -> Result<()> {
-    if dir_shares.is_empty() || supports_dir_shares {
-        return Ok(());
-    }
-    let shares = dir_shares
-        .iter()
-        .map(|s| format!("{}:{}", s.host_dir, s.guest_mount))
-        .collect::<Vec<_>>()
-        .join(", ");
-    anyhow::bail!(
-        "the {} backend cannot attach a host directory share, so --mount is \
-         unsupported here (requested: {shares}). Attach a disk-image volume \
-         instead (--volume HOST:GUEST:SIZE), or run on a backend with virtio-fs \
-         directory shares.",
-        backend_name,
-    )
-}
-
 pub fn resolve_launch(
     shape: &LaunchShape<'_>,
     admit: Option<&SessionAdmit<'_>>,
@@ -1532,47 +1504,6 @@ mod tests {
         );
         assert_eq!(hc.interval_secs, 10);
         assert_eq!(build_healthcheck(None, 30, 5, 3, 0), None);
-    }
-
-    /// `--mount` on a backend without virtio-fs shares is refused from the
-    /// arguments, and the message names the backend.
-    ///
-    /// Only HVF serves directory shares, so on every other backend this flag
-    /// used to be accepted, paid for — runtime detection, image resolution and
-    /// preparation, VM directory setup — and only then rejected inside the
-    /// runner. The refusal naming the *backend* is the part a caller can act
-    /// on; the runner's own message names the volume, which they already knew.
-    #[test]
-    fn a_dir_share_is_refused_up_front_on_a_backend_without_virtiofs() {
-        let shares = vec![DirShareSpec {
-            host_dir: "/host/fixtures".into(),
-            guest_mount: "/work/fixtures".into(),
-            read_only: true,
-        }];
-
-        let err = refuse_unsupported_dir_shares("firecracker", false, &shares)
-            .expect_err("a backend without virtio-fs shares must refuse --mount");
-        let msg = format!("{err:#}");
-        assert!(
-            msg.contains("firecracker"),
-            "the refusal must name the backend, got: {msg}"
-        );
-        assert!(
-            msg.contains("/work/fixtures"),
-            "the refusal must name the requested share, got: {msg}"
-        );
-
-        // The same request on a backend that serves shares is not refused.
-        refuse_unsupported_dir_shares("hvf", true, &shares)
-            .expect("a backend with virtio-fs shares must accept --mount");
-    }
-
-    /// No mounts, no refusal — the guard must not reject an ordinary launch,
-    /// including on a backend that cannot serve shares at all.
-    #[test]
-    fn a_launch_without_mounts_is_not_refused() {
-        refuse_unsupported_dir_shares("firecracker", false, &[])
-            .expect("a launch with no --mount must not be refused");
     }
 
     #[test]
